railz-api-client/pyproject.toml
[tool.poetry]
name = "railz-api-client"
version = "v2"
description = "A client library for accessing Railz API"
authors = []
readme = "README.md"
packages = [
    {include = "railz_api_client"},
]
include = ["CHANGELOG.md", "railz_api_client/py.typed"]


[tool.poetry.dependencies]
python = "^3.8"
httpx = ">=0.20.0,<0.28.0"
attrs = ">=21.3.0"
python-dateutil = "^2.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "I", "UP"]
